Add client certs to transport (#31)
This adds client certificates to the transport layer. If certificates are not
explicitly passed, then a self-signed certificate is generated using the DH key
passed as ClientConfig.KeyPair.

The self-signed certificate isn't really self-signed, since it's not a signing
key pair. It has a signature value of zero. Possession of the key is still
verified via triple DH.