Improve Reliable Tubes (#92)
Reliable tubes now close correctly using the FIN, ACK, FIN, ACK pattern from TCP.
Closing tubes ensures that all resources associated with the tube are freed.
Stopping the muxer now waits for all tubes to close properly or kills them automatically if they don't close.
Stopping a muxer also frees all resources associated with that hop session. No gorountines leak, and the underlying handle is closed.
We have a working fork of the nettests that works for unreliable connections. This fixes an intermittent bug in the tests for the transport layer as well.
The muxer now assigns tube IDs in sequential order with the parity of the tube indicating if it was opened by the client or the server.
The order of flags in tube headers now matches the order of flags in the paper.

Co-authored-by: baumanl <laurabauman12@gmail.com>