Add conditional compilation for trace level logging. Increase maximum packet size (#139)
One of our biggest performance losses is the fact that we are logging information about every single packet that a Hop connection receives. This fine grained level of logging is incredibly helpful for debugging, but it leads to unacceptable performance losses. This commit allows Hop to be compiled either with or without Trace level logging.

This is accomplished by using build tags. When Hop is built with the "debug" build tag, the variable `common.Debug` is set to true. It is false otherwise. Every call to `logrus.Trace` or equivalent function is now wrapped in an if statement checking if `common.Debug` is true.

This is not the best way to do this as it requires that every new call to `logrus.Trace` be wrapped in an  The other option would be to create our own logging wrapper that checks this condition. The problem is that part of the performance loss is caused by additional allocations needed do to the logging.